Check multiple paths for the shell script
Depending on how the library is installed, the shell script may be in different locations relative to the Validator.php class definition:

1. When this is the root package: ../vendor/bin/validate-htaccess
2. When this is installed in another package: ../../../bin/validate-htaccess

Additionally, non-standard Composer configurations could have the vendor dir in a completely different place. To support this, the path can be explicitly passed via the `HTACCESS_VALIDATOR_SCRIPT` environment variable.